Expose props to child components
I believe props are too important to be left behind. They do not
represent the state but they are often used to setup the
reusable components to better fit the context they are used in.

They are probably used for other sane things too, but the point is
that many react lib are using them so we can't just forget supporting
them.